simulator/requirements.txt
numpy
paho-mqtt
pandas
pytest
faker
jsonschema
requests